Introduce a top-level `LiteBox` type (#94)
This introduces a top-level `LiteBox` type to hold all the litebox-internal "globals" we might want to maintain. This is necessary for things such as upcoming shared FD storage (#31).

The members of `LiteBoxX` will be updated/added to as we proceed, but that should not change any public interfaces. This PR on the other hand _does_ change public interfaces (in particular, all the different subcomponents now require a `&LiteBox` to initialize, rather than a `&Platform`). This helps with #31, but would also help if/when we get the full version of #24 working. For now though, it unblocks a simpler variant of it, that I believe would be sufficient for #31.