Remove dependence on `std` for loader (#112)
The only usage of `fs` in `elf_loader` in our shim is for `IOError` part
of the enum. The actual choice of the enum doesn't _really_ matter for
this error, since it is only used to print out the message.

`elf_loader` itself doesn't _need_ to know if things are being `mmap`d
in, or they are being I/O'd in. Thus, by switching the error codes to be
an `MmapError`, we are able to remove `fs` from the required features
for `elf_loader`. This in turn allows us to remove the (transitive)
dependency both `libc` and `syscalls` within the shim (specifically,
`elf_loader` requires you to depend on at least one of the two, if you
have the `fs` feature around, even if you are not using them).