Reserve space for trampoline code and load trampoline for ld (#191)
This PR addresses two issues:

1. The elf loader we use would call `mmap` to load all segments with the 
`PT_LOAD` type at once and then invoke `mmap` again to change the memory 
protection. The trampoline code we map afterwards should be loaded right after
the executable file, but there is no guarantee that address is not taken already.
To address it, we need to allocate more memory when the first `mmap` is called. 
2. `ld` lib is also rewritten and hence we need to load its trampoline
code.