litebox_shim_linux: move file descriptor tables into `Task` (#464)
Eliminate `file_descriptors()` and `raw_descriptor_store()`, moving
their state into an `Arc` on `Task`. This eliminates some globals and
sets us up for supporting `!CLONE_FILES` in the future when desired.

Also, move the file system instance into an `Arc` that's referenced by
the `Task`. Update the pattern used to construct a Linux shim instance
to ensure the file system is set before the `Task`.

There are a few other globals to eliminate, but cutting this PR off here
to avoid it getting too big.