Add log_unsupported! macro and use it (#494)
Instead of panicking on unsupported functionality, invoke a new macro,
`log_unsupported!`, that prints to the debug log and continues,
typically by returning `EINVAL`. Update some (but not all) of the Linux
syscall handling path to follow this new pattern.

In the future, the macro can be updated to be configurable--to continue,
panic, or even print the stack of the user task or dump its core. For
now, the important part is to clearly tag the code paths that are known
to have unimplemented functionality, and to reduce subtle runtime panics
caused by the user process.