Optimize file system initialization (#563)
This PR optimizes our initialization sequence for `in_mem` file system,
which ends up improving our `node` runtime by cutting its initialization
time by more than half.

Essentially, a large amount of our initialization time was spent in the
initial `write` calls that were feeding in the data. So now there is a
new `initialize_primarily_read_heavy_file` function aimed purely for
initializing files that are meant to be read-heavy (such as program
data). It essentially postpones the cost of the memcpy to the time that
an actual `write` occurs on the file (which we expect to be never for
program data).

The other change is that for the runner, we load in the program data via
an `mmap` call rather than via a file read (similar to what we were
doing for tar files). This helps reduce that particular write too.

Along the way, I've also fixed a minor bug in
`litebox::fs::in_mem::FileSystem::truncate` when a longer size was
passed (previously it would just ignore it).

The benchmark summary from current `main` (summarized across 100
invocations):
| Benchmark | main (ms) | profitable-mola (ms) | Diff (ms) |

|:---------------------------|----------:|---------------------:|----------:|
| rewriter_hello_static | 14 | 14 | 0 |
| rewriter_node | 1055 | 1056 | 1 |
| run_rewritten_hello_static | 45 | 44 | -1 |
| run_rewritten_node | 242 | 109 | -133 |

(i.e., speeds up node invocation by ~55%).